Moved data for grid creation to its own file
