+++ src/grid.c
@@ -1,6 +1,7 @@
 #include <pebble.h>
 #include <stdlib.h>
 #include "grid.h"
+#include "gridData.h"
 
 int numberOfNeighbors(Grid grid, int cell);
 
@@ -111,159 +112,6 @@ int numberOfNeighbors(Grid grid, int cell) {
       count += grid.grid[cell - 1];
     
   return count;
-}
-
-Grid createBarberPole() {
-  int gridStride = 12;
-  int gridSize = gridStride * gridStride;
-  
-  //inital condition of game grid
-  char grid[] = 
-  {
-    0,0,0,0,0,0,0,0,0,0,1,1,
-    0,0,0,0,0,0,0,0,0,0,0,1,
-    0,0,0,0,0,0,0,0,0,1,0,0,
-    0,0,0,0,0,0,0,1,0,1,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,1,0,1,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,1,0,1,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,1,1,0,0,0,0,0,0,0,0,
-    1,0,0,0,0,0,0,0,0,0,0,0,
-    1,1,0,0,0,0,0,0,0,0,0,0,
-  };
-  //allocate heap space for array, and copy the initial data to the heap location
-  char* gridP = (char *)malloc(gridSize * sizeof(char));
-  memcpy(gridP, &grid, gridSize * sizeof(char));
-  
-  Grid gridStruct;
-  gridStruct.grid = gridP;
-  gridStruct.gridStride = gridStride;
-  gridStruct.gridSize = gridSize;
-  gridStruct.tempGrid = (char *)malloc(gridSize * sizeof(char));
-  return gridStruct;
-}
-
-Grid createQueenShuttle() {
-  int gridStride = 22;
-  int gridSize = gridStride * gridStride;
-  
-  char grid[] = {
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,1,1,0,0,0,1,0,0,1,0,0,0,0,0,0,0,0,0,0,1,1,
-    0,1,1,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,0,0,1,1,
-    0,0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  };
-  
-  //allocate heap space for array, and copy the initial data to the heap location
-  char* gridP = (char *)malloc(gridSize * sizeof(char));
-  memcpy(gridP, &grid, gridSize * sizeof(char));
-  
-  Grid gridStruct;
-  gridStruct.grid = gridP;
-  gridStruct.gridStride = gridStride;
-  gridStruct.gridSize = gridSize;
-  gridStruct.tempGrid = (char *)malloc(gridSize * sizeof(char));
-  return gridStruct;
-}
-
-Grid createMazing() {
-  int gridStride = 8;
-  int gridSize = gridStride * gridStride;
-  
-  //inital condition of game grid
-  char grid[] = 
-  {
-    0,0,0,0,0,0,0,0,
-    0,0,0,1,1,0,0,0,
-    0,1,0,1,0,0,0,0,
-    1,0,0,0,0,0,1,0,
-    0,1,0,0,0,1,1,0,
-    0,0,0,0,0,0,0,0,
-    0,0,0,1,0,1,0,0,
-    0,0,0,0,1,0,0,0,
-  };
-  //allocate heap space for array, and copy the initial data to the heap location
-  char* gridP = (char *)malloc(gridSize * sizeof(char));
-  memcpy(gridP, &grid, gridSize * sizeof(char));
-  
-  Grid gridStruct;
-  gridStruct.grid = gridP;
-  gridStruct.gridStride = gridStride;
-  gridStruct.gridSize = gridSize;
-  gridStruct.tempGrid = (char *)malloc(gridSize * sizeof(char));
-  return gridStruct;
-}
-
-Grid createPrePulsarShuttle() {
-  int gridStride = 30;
-  int gridSize = gridStride * gridStride;
-  
-  //inital condition of game grid
-  char grid[] = 
-  {
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,1,0,1,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,1,0,0,0,0,0,1,1,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,0,0,1,1,0,
-    0,0,1,0,0,0,0,0,1,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,
-    0,1,0,1,0,0,0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,1,0,0,
-    0,0,1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,1,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,1,1,0,0,0,1,1,1,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,1,0,0,0,0,0,1,0,1,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,
-    0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,
-  };
-  //allocate heap space for array, and copy the initial data to the heap location
-  char* gridP = (char *)malloc(gridSize * sizeof(char));
-  memcpy(gridP, &grid, gridSize * sizeof(char));
-  
-  Grid gridStruct;
-  gridStruct.grid = gridP;
-  gridStruct.gridStride = gridStride;
-  gridStruct.gridSize = gridSize;
-  gridStruct.tempGrid = (char *)malloc(gridSize * sizeof(char));
-  return gridStruct;
 }
 
 Grid getRandomGrid(Grid current) {
