+++ src/gridData.h
@@ -1,0 +1,10 @@
+#pragma once
+#include "grid.h"
+
+Grid createBarberPole();
+
+Grid createQueenShuttle();
+
+Grid createMazing();
+
+Grid createPrePulsarShuttle();
